requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
bcrypt
PyJWT